[MIG] sale_stock_prebook: Migration from 16.0
