+++ sale_stock_prebook/README.rst
@@ -16,14 +16,14 @@ sale_stock_prebook
 .. |badge2| image:: https://img.shields.io/badge/licence-LGPL--3-blue.png
     :target: http://www.gnu.org/licenses/lgpl-3.0-standalone.html
     :alt: License: LGPL-3
-.. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fsale--workflow-lightgray.png?logo=github
-    :target: https://github.com/OCA/sale-workflow/tree/14.0/sale_stock_prebook
-    :alt: OCA/sale-workflow
+.. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fsale--prebook-lightgray.png?logo=github
+    :target: https://github.com/OCA/sale-prebook/tree/16.0/sale_stock_prebook
+    :alt: OCA/sale-prebook
 .. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
-    :target: https://translation.odoo-community.org/projects/sale-workflow-14-0/sale-workflow-14-0-sale_stock_prebook
+    :target: https://translation.odoo-community.org/projects/sale-prebook-16-0/sale-prebook-16-0-sale_stock_prebook
     :alt: Translate me on Weblate
 .. |badge5| image:: https://img.shields.io/badge/runboat-Try%20me-875A7B.png
-    :target: https://runboat.odoo-community.org/builds?repo=OCA/sale-workflow&target_branch=14.0
+    :target: https://runboat.odoo-community.org/builds?repo=OCA/sale-prebook&target_branch=16.0
     :alt: Try me on Runboat
 
 |badge1| |badge2| |badge3| |badge4| |badge5|
@@ -46,10 +46,10 @@ The reservation creates the picking/moves via a procurement run by placing a rea
 Bug Tracker
 ===========
 
-Bugs are tracked on `GitHub Issues <https://github.com/OCA/sale-workflow/issues>`_.
+Bugs are tracked on `GitHub Issues <https://github.com/OCA/sale-prebook/issues>`_.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us to smash it by providing a detailed and welcomed
-`feedback <https://github.com/OCA/sale-workflow/issues/new?body=module:%20sale_stock_prebook%0Aversion:%2014.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+`feedback <https://github.com/OCA/sale-prebook/issues/new?body=module:%20sale_stock_prebook%0Aversion:%2016.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
 
 Do not contact contributors directly about support or help with technical issues.
 
@@ -88,6 +88,6 @@ Current `maintainer <https://odoo-community.org/page/maintainer-role>`__:
 
 |maintainer-mt-software-de| 
 
-This module is part of the `OCA/sale-workflow <https://github.com/OCA/sale-workflow/tree/14.0/sale_stock_prebook>`_ project on GitHub.
+This module is part of the `OCA/sale-prebook <https://github.com/OCA/sale-prebook/tree/16.0/sale_stock_prebook>`_ project on GitHub.
 
 You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
