+++ sale_stock_prebook/__manifest__.py
@@ -3,12 +3,12 @@
 {
     "name": "sale_stock_prebook",
     "summary": "Add process to prebook a sale order's stock before confirming it",
-    "version": "14.0.1.0.0",
+    "version": "16.0.1.0.0",
     "author": "MT Software, BCIM, Odoo Community Association (OCA)",
-    "website": "https://github.com/OCA/sale-workflow",
+    "website": "https://github.com/OCA/sale-prebook",
     "data": [
         "views/sale_views.xml",
-        "views/stock_location_route_views.xml",
+        "views/stock_route_views.xml",
     ],
     "depends": [
         "sale_stock",
