+++ sale_stock_prebook/models/__init__.py
@@ -1,5 +1,6 @@
-from . import sale
+from . import sale_order
+from . import sale_order_line
 from . import stock_move
 from . import stock_rule
 from . import report_stock_forecasted
-from . import stock_location_route
+from . import stock_route
