+++ sale_stock_prebook/models/stock_route.py
@@ -2,11 +2,10 @@
 from odoo import fields, models
 
 
-class Route(models.Model):
-    _inherit = "stock.location.route"
+class StockRoute(models.Model):
+    _inherit = "stock.route"
 
     no_sale_stock_prebook = fields.Boolean(
-        string="No Sale Stock Prebook",
         help="If set no stock will be prebooked, "
         "for configured Products with this route",
     )
