+++ setup/.setuptools-odoo-make-default-ignore
@@ -1,0 +1,2 @@
+# addons listed in this file are ignored by
+# setuptools-odoo-make-default (one addon per line)
