+++ setup/sale_stock_prebook/odoo/addons/sale_stock_prebook
@@ -1,0 +1,1 @@
+../../../../sale_stock_prebook
